+++ tux/modules/networking.py
@@ -982,8 +982,13 @@ class NetworkingPage(Adw.NavigationPage):
         
         # Header
         header = Adw.HeaderBar()
-        header.set_show_end_title_buttons(False)
-        header.set_show_start_title_buttons(False)
+        
+        # Back button
+        back_btn = Gtk.Button()
+        back_btn.set_icon_name("go-previous-symbolic")
+        back_btn.set_tooltip_text("Back to main menu")
+        back_btn.connect("clicked", lambda b: self.window.navigation_view.pop())
+        header.pack_start(back_btn)
         
         # Refresh button
         refresh_btn = Gtk.Button()