+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to Tux Assistant will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.7.16] - 2025-11-30
+
+### Fixed - Back Button on All Module Pages!
+- Added back arrow (←) button to ALL module pages
+- No more getting stuck in a module
+- Works on:
+  - Developer Tools
+  - Setup Tools
+  - Software Center
+  - Networking
+  - Desktop Enhancements
+  - Media Server
+  - Nextcloud
+  - Tux Tunes
+
 ## [5.7.15] - 2025-11-30
 
 ### Added - Documentation